fix: comentário javadoc

+++ src/main/java/com/trybe/java/regraprogressao/App.java
@@ -1,15 +1,20 @@
 package com.trybe.java.regraprogressao;
 
 import java.util.Scanner;
+/**
+ * App.
+ */
 
 public class App {
-
+  /**
+   * main.
+   */
   public static void main(String[] args) {
     Scanner scanner = new Scanner(System.in);
     System.out.println("Digite a quantidade de atividades para cadastrar: ");
     int quantity = Integer.parseInt(scanner.nextLine());
-    double weightSum = 0;
-    double totalAverage = 0;
+    int weightSum = 0;
+    int totalAverage = 0;
 
     for (int i = 1; i <= quantity; i++) {
       System.out.println("Digite o nome da atividade " + i + ": ");
